File storage module
The file storage module adds a persistent capability to the application by
serialising the instances created from BaseModel to a JSON file. When the
program starts up, the the file storage module also loads already
created and stored instances of the model, thereby having a list of all
instances created previously available.

The file storage module does not have a constructor because the values
needed, file_path and model instances variable are hard coded into the
class. The following describes the new additions and modifications:

FileStorage Class
This class main purpose is to serialise and deserialise model instances
to a JSON file.

Init file in the models package
In this file, we invoked the reload method of the FileStorage class to
make available to the programme model instances previously saved. So
anytime the programme starts up, this file is first looked up and the
reload method is invoked to access previously saved model instances

BaseModel Class
In order for the programme to automatically save and reload
previously saved model instances, the storage variable of the init file
inside of the models package is invoked in the constructor and the save
method of the BaseModel class.

Final note:
Currently debugging to fix some errors