Added the do_show method
This method prints a string representation of an instance of a model. It
is similar to the __str__() method of the BaseModel class which returns
a string representation of an instance of a model when invoked and can
be printed to the stdout with the print() method.

With the show method, we do not know the number of arguments that will
be passed upon invocation. We therefore use a string variable as the
parameter. We then split the values in the string variable in their
respective elements where spaces occur among them using the split()
method. This creates a list of the various elements and assign it to the
values variable as shown below.

values = params.split(" ")

We then obtain the name of the class and id using array subscripting and
we're certain that the class name is the first in the list at position
zero(0) and the id at position one(1). We then check for the
availability of each element for class name and id.

To check whether the class name is in the list of classes we have
created by creating a class variable of type dictionary with class name
as a member as demonstrated below. Here, objects are used as class names

objects = {"BaseModel": BaseModel}

If all checks pass, we combine the class name and id into a single value
separated by a period and pass this value as argument to the private
object method of the FileStorage class. The file storage class will then
use this information to fetch all the remaining data from the JSON file if
and only if the class id exists. If it doesn't, the error message, '**
no instance found **' is printed to the stdout.

Final note, I didn't check for pycodestyle compliance.