Added the destroy method
The destroy method implements the same approach as the show method. It
first splits the parameter using spaces as the delimiter and assigns
each element to variables of class name and class id.

It then proceeds with integrity checks to ensure it has been handed
a valid model instance, ensure the id exists among others. It then
creates a single value using the class name and id, here indicated as
object_name and object_id. It uses the system's built-in method, del()
and applies it on the all() method of the file storage class by
assigning the single value to it as demonstrated below:

del storage.all()[objs]

where storage is an instance of the FileStorage class, all() is a method
of the FileStorage class which returns all object instances and objs is
the combined value of object_name and object_id. del is a built-in
method in Python.

After deletion, it invokes the save() method of the FileStorage class to
update the JSON file with the changes.

Final note: passed pycodestyle compliance checks